Add `kit clone` command
This commit introduces the `kit clone` command
that permits to clone/sync one or more kits to a
specific directory.

The list of the kits is specified through a YAML
file in this format:

```yaml

kits:
- name: "core-kit"
  url: "https://github.com/macaroni-os/core-kit"
  branch: "mark-testing"
  #commit_sha1: "1456ea54c3efae418e8de419af2ea1b66ab2cbaa"
- name: "core-server-kit"
  url: "https://github.com/macaroni-os/core-server-kit"
  branch: "mark-testing"
```

where it's possible define the name of the kit,
the URL, optionally the branch and optionally a
specific hash.