`merge kit`: add --keep-workdir flag
By default is better to start from a fresh
status and for now is sound good to cleanup
the working directory at the end of the
elaboration.

This behaviour could be disable using
`--keep-workdir` flag.

In addition, I now avoid to run push
command if there aren't commits. This
avoid to generate errors.