Add `kit distfiles-sync` command!
Added the command to generate a local copy of
the distfiles directory from a list of kits.

It uses the same specfile used for merge command
with thirdpartymirrors in order to resolve the
ebuild urls with mirror scheme.

At the moment it's available only the local dir
backend but will be integrate S3 support in the
near future.

The FTP SRC_URI aren't yet supported.