`merge`: Correct commits with PRs
`go-git` keeps changes on checkout between
branches and this generate commits with
content merged between multiple PRs.

In order to avoid this issue it's better
reset the files committed to the original
states and reset them from worktree.

This commit fix the behaviour of PR
for atoms, eclasses and fixups.