`merge`: Automatically bump new revision for changes
If a specific version ebuild has changes now
we increment the revision on target kit.

NOTE: At the moment, this feature is not complete.
If a packages is already been bumped with a new revision
changes related to the second changes are not catched.
Considering, that changes of existing packages are not
numerous catch the first change could be sufficient.
The users that want force a new revision could force
this from the source kit with a revision bigger of the
existing or directly create manually a new PR.