Add `builtin_dirlisting` generator
The `builtin_dirlisting` generator permits
to generate packages reading an HTTP indexes page.
After read the href links of the package matching
the regex string defined in the `matcher` field
it's needed using transform in order to obtain the
list of versions availables.
So, if the naming convention of a specific file is
`foo-<version>.tar.gz` the trasform must be used to
remove the prefix `foo-` and the suffix `.tar.gz`.

The list of the versions are sort in reverse order
to have the major version as first element of the
array.

In addition, this commit add support to `vars` attribute
correctly, that permits to set custom variables pushed
in the map used in the render phase.

An example:

```
packages:
  - foo:
      vars:
        desc: My foo app
```