autogen: Review releases support
It's possible that an old release
has issues on retrieve tag metadata
using the ListTags function. For example,
there are more tags that releases.
In order to avoid segfault, i skip
releases without tag metadata.

We prefer avoid iterate for every release and
executing a specific call to retrieve tag
metadata of the version.