extensions/rust: Review management of git crates
The git crates of a package are now managed
with a map because it possible that the same
repository supplies multiple dependencies.
In this case the file mark_config.toml must
be created with the definition of all packages
available.

For example the package `fish` is an example
with this use case.