generator/builtin-github: Add `ignore_tags` attribute
The normal implementation using query
`releases` retrive the release information
and then use the tag name to retrieve the
tag object and retrieve different informations.

This is not safe for some weird projects
using github releases in a anomal way, for
example the tag name inside the release object
doesn't reference the real tag name, etc.

In this case, it's better just using the release
object and the field `TargetCommitish`.

In order to use this new way it's needed
add the new attribute `ignore_tags`:

```
github:
  query: releases
  ignore_tags: true
```

The default value is false and it used
the previous implementation.