generators/builtin-json: Parse map[key]interfaces for versions
In order to parsing file where the version is
defined as string of the direct object returned
by the URL, for example:

{
   "0.1.0": { "tarball": "0.1.0.tar.gz" },
   "0.2.0": { "tarball": "0.2.0.tar.gz" }
}

we need to parse the object directly from the
code because doesn't exist a `keys` macro on
ohler55/ojg library.

So, in this case we need to get the full object
with a filter using value '$' and the generator
automatically will use all the keys as versions
if the object is a `map[string]interface`.